Ensure field is active and has a result before checking for errors.
Fields delegate error related methods to their result object. If a field
has no result, calling methods such as 'has_errors' dies with a Moose
delegation.

Check the field is active and has a result before checking if it has
errors. This is the same approach taken in
  HTML::FormHandler::Fields::_fields_validate
before deferring to the fields validate function.